feat: track stats info
