+++ p2p_wallet/Scenes/Main/ReceiveToken/NetworkSelection/ConfirmReceivingBitcoin/RenBTCStatusService/RenBTCStatusService.swift
@@ -96,7 +96,12 @@ class RenBTCStatusService: RenBTCStatusServiceType {
                         return self.feeRelayer
                             .topUpAndRelayTransaction(
                                 preparedTransaction: preparedTransaction,
-                                payingFeeToken: .init(address: payingFeeAddress, mint: payingFeeMintAddress)
+                                payingFeeToken: .init(
+                                    address: payingFeeAddress,
+                                    mint: payingFeeMintAddress
+                                ),
+                                operationType: .other("CreateAssociatedToken"), // TODO: - Which?
+                                currency: SolanaSDK.PublicKey.renBTCMint.base58EncodedString
                             )
                             .map(\.first)
                     }
